Convert server to use floats

+++ src/main/kotlin/ca/joelathiessen/kaly2/Measurement.kt
@@ -2,4 +2,4 @@ package ca.joelathiessen.kaly2
 
 import lejos.robotics.navigation.Pose
 
-class Measurement(val distance: Double, val angle: Double, val pose: Pose, val time: Long)
+class Measurement(val distance: Float, val angle: Float, val pose: Pose, val time: Long)
