+++ src/main/kotlin/ca/joelathiessen/kaly2/planner/PathSegment.kt
@@ -1,12 +1,12 @@
 package ca.joelathiessen.kaly2.planner
 
-abstract class PathSegment(val x: Double, val y: Double, parentIn: PathSegment?, costIn: Double) : PathSegmentInfo {
-    var cost: Double = costIn
+abstract class PathSegment(val x: Float, val y: Float, parentIn: PathSegment?, costIn: Float) : PathSegmentInfo {
+    var cost: Float = costIn
         protected set
 
     var parent: PathSegment? = parentIn
         protected set
 
-    abstract fun makeChild(xChild: Double, yChild: Double): PathSegment?
+    abstract fun makeChild(xChild: Float, yChild: Float): PathSegment?
     abstract fun changeParentIfCheaper(posNewParent: PathSegment)
 }
