+++ src/main/kotlin/ca/joelathiessen/kaly2/planner/linear/LinearPathSegmentRootFactory.kt
@@ -6,6 +6,6 @@ import ca.joelathiessen.kaly2.planner.PathSegmentRootFactory
 
 class LinearPathSegmentRootFactory : PathSegmentRootFactory {
     override fun makePathSegmentRoot(startPose: RobotPose): PathSegment {
-        return LinearPathSegment(startPose.x.toDouble(), startPose.y.toDouble(), null, 0.0)
+        return LinearPathSegment(startPose.x, startPose.y, null, 0.0f)
     }
 }
