+++ src/main/kotlin/ca/joelathiessen/kaly2/subconscious/Subconscious.kt
@@ -20,8 +20,8 @@ class Subconscious(private val sensor: Kaly2Sensor, private val pilot: Different
     override fun run() {
         println("Subconscious starting")
 
-        var distance: Double// distance to detected point in m
-        var angle: Double// angle to detected point in radians
+        var distance: Float// distance to detected point in m
+        var angle: Float// angle to detected point in radians
         var time: Long
         var pose: Pose
 
@@ -51,7 +51,7 @@ class Subconscious(private val sensor: Kaly2Sensor, private val pilot: Different
 
                 sensor.fetchSample(sensorReading, 0)
 
-                distance = sensorReading[0].toDouble()
+                distance = sensorReading[0]
                 angle = spinner.angle
                 pose = odometry.pose
                 time = System.currentTimeMillis()
